bugfix: testbed - create java objects dir if necessary

+++ testDriver/src/com/bbn/rpki/test/objects/TestbedConfig.java
@@ -62,7 +62,12 @@ public class TestbedConfig implements Constants {
    * @param iniFile
    */
   public TestbedConfig(String iniFile) {
-    Util.deleteDirectories(new File(OBJECT_PATH).listFiles());
+    File objectsDir = new File(OBJECT_PATH);
+    if (!objectsDir.exists()) {
+      objectsDir.mkdir();
+    } else {
+      Util.deleteDirectories(new File(OBJECT_PATH).listFiles());
+    }
     try {
       wini = new Wini(new StringReader(iniFile));
       Collection<Map.Entry<String, Section>> sectionEntries = wini.entrySet();
